requirements.tech-report.txt
apache-beam[gcp]==2.43.0
google-cloud-firestore==2.13.0
